Use u8 instead of u16 for bitflags
It is wide enough

+++ src/system/molecules.rs
@@ -114,7 +114,7 @@ mod connect {
         /// The `Connectivity` bitflag encode the topological distance between
         /// two particles in the molecule, i.e. the number of bonds between the
         /// particles.
-        flags Connectivity: u16 {
+        flags Connectivity: u8 {
             /// The particles are separated by one bond
             const CONNECT_12   = 0b0001,
             /// The particles are separated by two bonds
